agents_service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic>=2.12.2
python-dotenv==1.0.1
openai-agents==0.14.2